@@ -5,40 +5,42 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.6.0 <4.0.0'
+
+resolution: workspace
 
 dependencies:
   flutter:
     sdk: flutter
-  go_router: ^14.2.0
-  flutter_riverpod: ^2.3.6
-  riverpod_annotation: ^2.1.1
+  go_router: ^16.2.1
+  flutter_riverpod: ^3.0.0-dev.17
+  riverpod_annotation: ^3.0.0-dev.17
   material_design_icons_flutter: ^7.0.7296
   shared:
     path: ../packages/shared
   dio: ^5.1.1
   logger: ^2.0.2+1
-  firebase_crashlytics: ^4.0.2
-  firebase_core: ^3.1.1
+  firebase_crashlytics: ^5.0.1
+  firebase_core: ^4.1.0
   intl: any
   flutter_localizations:
     sdk: flutter
-  firebase_auth: ^5.1.1
-  google_sign_in: ^6.1.0
-  fl_chart: ^0.70.2
+  firebase_auth: ^6.0.2
+  google_sign_in: ^7.1.1
+  fl_chart: ^1.1.0
   sliver_tools: ^0.2.10
-  flutter_dotenv: ^5.0.2
-  freezed_annotation: ^2.4.2
+  flutter_dotenv: ^6.0.0
+  freezed_annotation: ^3.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^5.0.0
-  custom_lint: ^0.7.1
-  riverpod_lint: ^2.3.7
+  flutter_lints: ^6.0.0
+  custom_lint: ^0.8.0
+  riverpod_lint: ^3.0.0-dev.17
   build_runner: ^2.3.3
-  riverpod_generator: ^2.2.1
-  freezed: ^2.5.2
+  riverpod_generator: ^3.0.0-dev.17
+  freezed: ^3.2.0
 
 
 flutter:
