@@ -3,15 +3,17 @@ description: A starting point for Dart libraries or applications.
 version: 1.0.0
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.6.0 <4.0.0'
+
+resolution: workspace
 
 dependencies:
-  freezed_annotation: ^2.2.0
+  freezed_annotation: ^3.1.0
   json_annotation: ^4.8.1
 
 dev_dependencies:
   build_runner: ^2.3.3
-  freezed: ^2.5.8
+  freezed: ^3.2.0
   json_serializable: ^6.6.2
-  lints: ^5.1.1
+  lints: ^6.0.0
   test: ^1.21.0
