@@ -1,7 +1,18 @@
 name: expenses_tracker
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.6.0 <4.0.0'
+
+workspace:
+  - client
+  - server
+  - packages/shared
 
 dev_dependencies:
-  melos: ^3.0.1
+  melos: ^7.1.0
+
+scripts:
+  generate:
+    name: generate
+    run: melos exec -c 1 --depends-on="build_runner" --flutter -- "flutter pub run build_runner build --delete-conflicting-outputs"
+    description: Build all generated files for Flutter packages in this project.
