@@ -4,14 +4,16 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.6.0 <4.0.0'
+
+resolution: workspace
 
 dependencies:
   basic_utils: ^5.5.4
   dart_frog: ^1.0.1
-  dart_jsonwebtoken: ^2.7.1
+  dart_jsonwebtoken: ^3.2.0
   dotenv: ^4.1.0
-  freezed_annotation: ^2.2.0
+  freezed_annotation: ^3.1.0
   http: ^1.1.2
   json_annotation: ^4.8.1
   postgres: ^3.0.5
@@ -20,11 +22,11 @@ dependencies:
 
 dev_dependencies:
   build_runner: ^2.3.3
-  freezed: ^2.3.3
+  freezed: ^3.2.0
   json_serializable: ^6.6.2
   mocktail: ^1.0.2
   test: ^1.19.2
-  very_good_analysis: ^7.0.0
+  very_good_analysis: ^9.0.0
 
 assets:
   - .env
