@@ -6,21 +6,46 @@ part of 'error_service.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$errorServiceHash() => r'7157ae0e0c85ce36de7017bf34b00a57e7178638';
-
-/// See also [errorService].
 @ProviderFor(errorService)
-final errorServiceProvider = AutoDisposeProvider<ErrorService>.internal(
-  errorService,
-  name: r'errorServiceProvider',
-  debugGetCreateSourceHash:
-      const bool.fromEnvironment('dart.vm.product') ? null : _$errorServiceHash,
-  dependencies: null,
-  allTransitiveDependencies: null,
-);
+const errorServiceProvider = ErrorServiceProvider._();
 
-@Deprecated('Will be removed in 3.0. Use Ref instead')
-// ignore: unused_element
-typedef ErrorServiceRef = AutoDisposeProviderRef<ErrorService>;
+final class ErrorServiceProvider
+    extends $FunctionalProvider<ErrorService, ErrorService, ErrorService>
+    with $Provider<ErrorService> {
+  const ErrorServiceProvider._()
+      : super(
+          from: null,
+          argument: null,
+          retry: null,
+          name: r'errorServiceProvider',
+          isAutoDispose: true,
+          dependencies: null,
+          $allTransitiveDependencies: null,
+        );
+
+  @override
+  String debugGetCreateSourceHash() => _$errorServiceHash();
+
+  @$internal
+  @override
+  $ProviderElement<ErrorService> $createElement($ProviderPointer pointer) =>
+      $ProviderElement(pointer);
+
+  @override
+  ErrorService create(Ref ref) {
+    return errorService(ref);
+  }
+
+  /// {@macro riverpod.override_with_value}
+  Override overrideWithValue(ErrorService value) {
+    return $ProviderOverride(
+      origin: this,
+      providerOverride: $SyncValueProvider<ErrorService>(value),
+    );
+  }
+}
+
+String _$errorServiceHash() => r'd5fa0ea2b5bc45f9a74531d1aebb6334ed76e8d9';
+
 // ignore_for_file: type=lint
 // ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package
