@@ -6,21 +6,39 @@ part of 'categories.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$categoriesHash() => r'2edb27866644eb8d34a4c1da8032fedef026077c';
-
-/// See also [categories].
 @ProviderFor(categories)
-final categoriesProvider = AutoDisposeFutureProvider<List<Category>>.internal(
-  categories,
-  name: r'categoriesProvider',
-  debugGetCreateSourceHash:
-      const bool.fromEnvironment('dart.vm.product') ? null : _$categoriesHash,
-  dependencies: null,
-  allTransitiveDependencies: null,
-);
+const categoriesProvider = CategoriesProvider._();
 
-@Deprecated('Will be removed in 3.0. Use Ref instead')
-// ignore: unused_element
-typedef CategoriesRef = AutoDisposeFutureProviderRef<List<Category>>;
+final class CategoriesProvider extends $FunctionalProvider<
+        AsyncValue<List<Category>>, List<Category>, FutureOr<List<Category>>>
+    with $FutureModifier<List<Category>>, $FutureProvider<List<Category>> {
+  const CategoriesProvider._()
+      : super(
+          from: null,
+          argument: null,
+          retry: null,
+          name: r'categoriesProvider',
+          isAutoDispose: true,
+          dependencies: null,
+          $allTransitiveDependencies: null,
+        );
+
+  @override
+  String debugGetCreateSourceHash() => _$categoriesHash();
+
+  @$internal
+  @override
+  $FutureProviderElement<List<Category>> $createElement(
+          $ProviderPointer pointer) =>
+      $FutureProviderElement(pointer);
+
+  @override
+  FutureOr<List<Category>> create(Ref ref) {
+    return categories(ref);
+  }
+}
+
+String _$categoriesHash() => r'af0d71ae83325984429ecf424a45f1882778d093';
+
 // ignore_for_file: type=lint
 // ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package
