@@ -1,14 +1,13 @@
 import 'package:dio/dio.dart';
 import 'package:expensetrack/core/services/dio.dart';
 import 'package:expensetrack/core/services/error_service.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:shared/shared.dart';
 
 part 'categories_repository.g.dart';
 
 @riverpod
-CategoriesRepository categoriesRepository(CategoriesRepositoryRef ref) {
+CategoriesRepository categoriesRepository(Ref ref) {
   return CategoriesRepository(ref);
 }
 
